update denom config
